Brings in the decomp work that lands on the sync's compile and link walls: the generated headers' size assertions guarded off the host build (#2660), the six model headers' host destructor split and the nine size_t operator new headers (#2663), Player.cpp compiling on the host (#2666), the declaration return-type rows (#2669), two re-aligned hand-assembled TUs (#2672) and the day's matches and renames.

Eleven files conflicted and every one took main's side: contributions.json, the six model headers where main's version is this branch's plus the non-virtual destructor declaration, include/dScMgPachinko_c.h and include/dScMgSmartball_c.h where the two sides differ only in line endings, src/func_0205d23c.c and tools/asset_catalog.py where main's text is the rewritten comment. No port-only code sits in any of them; the only port-authored text lost is a comment paragraph in include/ModelBase.h that main's own rewrite of that comment dropped.

+++ include/daObjFl_Gura_c.h
@@ -47,14 +47,14 @@ struct daObjFl_Gura_c : daObjGuragura_c {
     int CleanupResources();            /* slot  3 */
     int InitResources();               /* slot  0 */
 
-    static void *operator new(unsigned long size);
+    static void *operator new(size_t size);
 
     /* Declared last and inline so class instantiation emits the retail D1/D0
        pair in cartridge order without a separate leaf D2 body. */
     virtual ~daObjFl_Gura_c() {}
 };
 
-inline void *daObjFl_Gura_c::operator new(unsigned long size)
+inline void *daObjFl_Gura_c::operator new(size_t size)
 {
     return _ZN7fBase_cnwEj((unsigned)size);
 }
