+++ include/daObjKi_Ita_c.h
@@ -30,14 +30,14 @@ extern "C" void *_ZN7fBase_cnwEj(unsigned size);
 struct daObjKi_Ita_c : daObjFloatBoard_c {
     int InitResources();               /* slot  0 */
 
-    static void *operator new(unsigned long size);
+    static void *operator new(size_t size);
 
     /* Declared last and inline so class instantiation can emit the retail
        D1/D0 pair in cartridge order without a separate D2 body. */
     virtual ~daObjKi_Ita_c() {}
 };
 
-inline void *daObjKi_Ita_c::operator new(unsigned long size)
+inline void *daObjKi_Ita_c::operator new(size_t size)
 {
     return _ZN7fBase_cnwEj((unsigned)size);
 }
