+++ include/daObjKm1_Kuruma_c.h
@@ -30,14 +30,14 @@ struct daObjKm1_Kuruma_c : daObjKuruma_c {
     int CleanupResources();            /* slot  3 */
     int InitResources();               /* slot  0 */
 
-    static void *operator new(unsigned long size);
+    static void *operator new(size_t size);
 
     /* MEASURED -- INLINE ON PURPOSE. Declaring this last makes the class TU
        emit D1 before D0, in the cartridge's order, without a leaf D2 body. */
     virtual ~daObjKm1_Kuruma_c() {}    /* slots 16 (D1), 17 (D0) */
 };
 
-inline void *daObjKm1_Kuruma_c::operator new(unsigned long size)
+inline void *daObjKm1_Kuruma_c::operator new(size_t size)
 {
     return _ZN7fBase_cnwEj((unsigned)size);
 }
